implementation of PR review comments
